+++ tests/testthat/test-rev_niveles.R
@@ -1,27 +1,35 @@
-var_labelled <- structure(c(1, 2, 2, 9),
-                          labels = c('a' = 1, 'b' = 2, 'nr' = 9),
+var_labelled <- structure(c(1, 2, 2, 3, 9),
+                          labels = c('a' = 1, 'b' = 2, 'c' = 3, 'nr' = 9),
                           label = 'this is labelled',
                           class = 'haven_labelled')
 
 test_that("rev_niveles haven_labelled", {
 
-  # Valores
-  expect_equal(as.numeric(rev_niveles(var_labelled, rango_inv = NULL)), c(9, 8, 8, 1))
-  # Etiquetas
-  expect_equal(attr(rev_niveles(var_labelled, rango_inv = NULL), 'labels'), c('a' = 9, 'b' = 8, 'nr' = 1))
+  # Valores sin rango
+  expect_equal(as.numeric(rev_niveles(var_labelled, rango_inv = NULL)), c(9, 8, 8, 7, 1))
+  # Etiquetas sin rango
+  expect_equal(attr(rev_niveles(var_labelled, rango_inv = NULL), 'labels'), c('a' = 9, 'b' = 8, 'c' = 7, 'nr' = 1))
 
-  # Valores
-  expect_equal(as.numeric(rev_niveles(var_labelled, rango_inv = c(1, 2))), c(2, 1, 1, 9))
-  # Etiquetas
-  expect_equal(attr(rev_niveles(var_labelled, rango_inv = c(1, 2)), 'labels'), c('a' = 2, 'b' = 1, 'nr' = 9))
+  # Valores con rango
+  expect_equal(as.numeric(rev_niveles(var_labelled, rango_inv = c(1, 2))), c(2, 1, 1, 3, 9))
+  # Etiquetas con rango
+  expect_equal(attr(rev_niveles(var_labelled, rango_inv = c(1, 2)), 'labels'), c('a' = 2, 'b' = 1, 'c' = 3, 'nr' = 9))
+
+  # Igualdad entre rango == NULL y todo el rango
+  expect_equal(rev_niveles(var_labelled, rango_inv = NULL),
+               rev_niveles(var_labelled, rango_inv = c(1, 9)))
 })
 
 test_that("rev_niveles factor", {
-  var_factor <- factor(c('a', 'b', 'b', 'nr'))
+  var_factor <- factor(c('a', 'b', 'b', 'c', 'nr'))
 
-  expect_equal(levels(rev_niveles(var_factor, niveles_inv = c('a', 'b'))), c('b', 'a', 'nr'))
+  expect_equal(levels(rev_niveles(var_factor, niveles_inv = c('a', 'b'))), c('b', 'a', 'c', 'nr'))
 
-  expect_equal(levels(rev_niveles(var_factor)), c('nr', 'b', 'a'))
+  expect_equal(levels(rev_niveles(var_factor)), c('nr', 'c', 'b', 'a'))
+
+  # Igualdad entre rango == NULL y todo el rango
+  expect_equal(rev_niveles(var_factor, niveles_inv = NULL),
+               rev_niveles(var_factor, niveles_inv = levels(var_factor)))
 })
 
 test_that("rev_niveles numeric", {
@@ -33,5 +41,9 @@ test_that("rev_niveles numeric", {
 
   expect_equal(rev_niveles(var_numeric, rango_inv = c(1, 2)), c(2, 1, 1, 9))
 
-  expect_equal(as.numeric(rev_niveles(var_labelled, rango_inv = c(1, 2))), c(2, 1, 1, 9))
+  expect_equal(as.numeric(rev_niveles(var_labelled, rango_inv = c(1, 2))), c(2, 1, 1, 3, 9))
+
+  # Igualdad entre rango == NULL y todo el rango
+  expect_equal(rev_niveles(var_numeric, rango_inv = NULL),
+               rev_niveles(var_numeric, rango_inv = c(1, 9)))
 })